# AI-Assignments/realTime-foodDelivery-service/API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: SwiftEats API
  version: 1.0.0
  description: |
    API specification for **SwiftEats** — a real-time food delivery platform.

    ## Architecture

    Six Spring Boot microservices share domain code in `platform-lib`. **All external clients**
    (UIs, Postman, curl) call **backend-service** at `http://localhost:8080`. The gateway
    proxies domain routes to internal services.

    | Service | Port | Routes |
    |---------|------|--------|
    | backend-service | 8080 | `/auth/*`, public `/restaurants/*`, `/drivers/*`, order tracking |
    | entities-service | 8081 | `/admin/restaurants/*` |
    | order-service | 8082 | `/orders/*`, `/admin/orders/*` |
    | payment-service | 8083 | `/payments/*` |
    | refund-service | 8084 | `/refunds/*` |
    | analytics-service | 8085 | `/analytics/*`, `/admin/analytics/*` |

    **Unified Swagger UI:** http://localhost:8080/swagger-ui.html (dropdown for each service spec).

    ## Modules

    - **Auth** — customer register, login, profile (backend-service)
    - **Restaurant** — browse, search, menu cache (backend); admin CRUD (entities-service)
    - **Order** — idempotent creation, async payment, state machine (order-service)
    - **Payment** — RabbitMQ worker, mock gateway (payment-service)
    - **Refund** — initiation → successful mock refund (refund-service)
    - **Tracking** — Kafka GPS ingest, Redis hot path, SSE (backend-service)
...

    **Order state flow:**
...

...

...

...

...

...

  contact:
...
  license:
...
servers:
...
paths:
...
  /api/v1/restaurants:
    get:
      tags: [Restaurants]
      summary: Search and list restaurants
      operationId: listRestaurants
      parameters:
        - name: city
          in: query
          schema: { type: string }
        - name: cuisine
          in: query
          schema: { type: string }
        - name: isOpen
          in: query
          schema: { type: boolean }
        - name: minRating
          in: query
          schema: { type: number, format: double }
        - name: name
          in: query
          description: Partial name search
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated restaurant list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantPageResponse'
  /api/v1/restaurants/{id}/menu:
    get:
      tags: [Restaurants]
      summary: Get restaurant menu and status (P99 critical path)
      operationId: getRestaurantMenu
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
      responses:
        '200':
          description: Menu bundle with restaurant status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantMenuResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/restaurants:
    get:
      tags: [Admin Restaurants]
      summary: List restaurants for admin
      description: Served by **entities-service** (proxied via backend gateway).
      operationId: listAdminRestaurants
      security:
        - AdminApiKey: []
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [PENDING, ACTIVE, SUSPENDED] }
          description: Optional filter by restaurant status
      responses:
        '200':
          description: Restaurant summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminRestaurantSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Admin Restaurants]
      summary: Onboard a new restaurant
      operationId: createRestaurant
      security:
        - AdminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRestaurantRequest'
      responses:
        '201':
          description: Restaurant created (status PENDING)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/admin/restaurants/{id}:
    get:
      tags: [Admin Restaurants]
      summary: Get restaurant detail (admin)
      operationId: getAdminRestaurant
      security:
        - AdminApiKey: []
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
      responses:
        '200':
          description: Restaurant with menu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Admin Restaurants]
      summary: Update restaurant profile or status
      operationId: updateRestaurant
      security:
        - AdminApiKey: []
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRestaurantRequest'
      responses:
        '200':
          description: Updated restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin Restaurants]
      summary: Soft-delete restaurant (status → SUSPENDED)
      operationId: deleteRestaurant
      security:
        - AdminApiKey: []
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
      responses:
        '204':
          description: Restaurant suspended
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/restaurants/{id}/approve:
    patch:
      tags: [Admin Restaurants]
      summary: Approve restaurant (ACTIVE + open)
      operationId: approveRestaurant
      security:
        - AdminApiKey: []
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
      responses:
        '200':
          description: Approved restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/restaurants/{id}/menu-items:
    post:
      tags: [Admin Restaurants]
      summary: Add menu item
      operationId: addMenuItem
      security:
        - AdminApiKey: []
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuItemRequest'
      responses:
        '201':
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/restaurants/{id}/menu-items/{itemId}:
    patch:
      tags: [Admin Restaurants]
      summary: Update menu item
      operationId: updateMenuItem
      security:
        - AdminApiKey: []
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - $ref: '#/components/parameters/MenuItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMenuItemRequest'
      responses:
        '200':
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin Restaurants]
      summary: Delete menu item
      operationId: deleteMenuItem
      security:
        - AdminApiKey: []
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - $ref: '#/components/parameters/MenuItemId'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Api-Key
    CustomerIdAuth:
      type: apiKey
      in: header
      name: X-Customer-Id
      description: Customer UUID; required together with X-Customer-Api-Key on protected routes
    CustomerApiKeyAuth:
      type: apiKey
      in: header
      name: X-Customer-Api-Key
      description: Per-customer API token from the customer record; required together with X-Customer-Id
    DriverAuth:
      type: apiKey
      in: header
      name: X-Driver-Api-Key
      description: Must match the driver UUID in the request path
  parameters:
    RestaurantId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    MenuItemId:
      name: itemId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, timestamp, path]
      properties:
        code: { type: string, example: NOT_FOUND }
        message: { type: string }
        timestamp: { type: string, format: date-time }
        path: { type: string }
...
    MenuItemResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name:
          type: string
          example: "[T] Kolhapuri Misal"
      type: object
      properties:
        restaurantId: { type: string, format: uuid }
        name:
          type: string
          example: "[T] Misal House"
        estimatedWaitMins: { type: integer }
        menuItems:
          type: array
          items:
            $ref: '#/components/schemas/MenuItemResponse'
        cachedAt: { type: string, format: date-time }
    CreateRestaurantRequest:
      type: object
      required: [name, address, city, cuisines]
      properties:
        name:
          type: string
          description: Stored with `[T] ` prefix (e.g. admin sends `Pune Biryani Co`, persisted as `[T] Pune Biryani Co`)
          example: Pune Biryani Co
        address: { type: string }
        city: { type: string }
        cuisines:
          type: array
          items: { type: string }
        contactEmail: { type: string, format: email }
        openingTime: { type: string, example: "10:00" }
        closingTime: { type: string, example: "23:00" }
    RestaurantDetailResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name:
          type: string
          example: "[T] Pune Biryani Co"
        cuisines:
          type: array
          items: { type: string }
        contactEmail: { type: string }
        openingTime: { type: string }
        closingTime: { type: string }
        estimatedWaitMins: { type: integer }
        createdAt: { type: string, format: date-time }
    AdminRestaurantSummaryResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        city: { type: string }
        status:
          type: string
          enum: [PENDING, ACTIVE, SUSPENDED]
        isOpen: { type: boolean }
    UpdateRestaurantRequest:
      type: object
      properties:
        name: { type: string }
        address: { type: string }
        city: { type: string }
        cuisines:
          type: array
          items: { type: string }
        contactEmail: { type: string, format: email }
        openingTime: { type: string, example: "10:00" }
        closingTime: { type: string, example: "23:00" }
        isOpen: { type: boolean }
        estimatedWaitMins: { type: integer }
    CreateMenuItemRequest:
      type: object
      required: [name, category, price]
      properties:
        name: { type: string }
        description: { type: string }
        category: { type: string }
        price: { type: number, format: double, minimum: 0.01 }
        available: { type: boolean, default: true }
    UpdateMenuItemRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        category: { type: string }
        price: { type: number, format: double, minimum: 0.01 }
        available: { type: boolean }
...